sdk/python/pyproject.toml
[tool.poetry]
name = "virtengine"
version = "0.1.0"
description = "Python SDK for VirtEngine blockchain"
authors = ["VirtEngine <sdk@virtengine.io>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/virtengine/virtengine"
repository = "https://github.com/virtengine/virtengine/tree/main/sdk/python"
documentation = "https://docs.virtengine.io/sdk/python"
keywords = ["blockchain", "cosmos", "sdk", "identity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "virtengine" },
    { include = "cosmos" },
    { include = "cosmos_proto" },
    { include = "gogoproto" },
    { include = "google" },
    { include = "amino" },
    { include = "k8s_io" },
]
include = ["virtengine/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.60.0"
protobuf = "^4.25.0"
bip-utils = "^2.9.0"
ecdsa = "^0.18.0"
bech32 = "^1.2.0"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
